requirements.txt
openai>=1.0.0
chromadb
casbin
streamlit
python-dotenv
pandas
matplotlib
pysqlite3-binary